+++ mttl/models/containers/selectors/base_selectors.py
@@ -10,12 +10,11 @@ from pyparsing import abstractmethod
 from torch import nn
 from torch.distributions import Categorical
 
-from mttl.logging import logger
+from mttl.logging import logger, warn_once
 from mttl.models.library.expert import ExpertInfo
 from mttl.models.ranker.adapter_ranker import AdapterRankerHelper
 from mttl.models.ranker.classifier_ranker import ClusterPredictor
 from mttl.models.utils import MetricLogger
-from mttl.utils import warn_once
 
 SELECTORS_NAME_TO_KLASS = {}
 SELECTORS_CONFIG_TO_NAME = {}
