+++ mttl/models/modifiers/lora.py
@@ -9,9 +9,9 @@ import numpy as np
 import torch
 from torch import nn
 
+from mttl.logging import warn_once
 from mttl.models.modifiers import register_modifier
 from mttl.models.modifiers.base import MergeableAdapter, ModifierConfig, ModifyMixin
-from mttl.utils import warn_once
 
 
 @dataclass
